YAK SHAVING REFACTOR
* major: restructure warehouses into module and add a file per warehouse e.g. bigquery
* minor: remove some comments
* minor: remove Dash - doesn't appear to be used
* minor: small cleanup of some syntax
