+++ cli/src/warehouse/constants.rs
@@ -1,0 +1,1 @@
+pub const GOOGLE_ENV_VAR: &str = "GOOGLE_APPLICATION_CREDENTIALS";
