+++ cli/src/warehouse/errors.rs
@@ -1,0 +1,10 @@
+use std::fmt;
+
+#[derive(Debug, Clone)]
+pub struct ParseMetadataSourceError {}
+
+impl fmt::Display for ParseMetadataSourceError {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "Invalid metadata source.")
+    }
+}
